+++ app/src/main/java/org/openkis/android/data/local/AppDatabase.kt
@@ -6,19 +6,22 @@ import org.openkis.android.data.local.dao.ArtificialDao
 import org.openkis.android.data.local.dao.CaveDao
 import org.openkis.android.data.local.dao.ServerDao
 import org.openkis.android.data.local.dao.SpringDao
+import org.openkis.android.data.local.dao.SurveyDao
 import org.openkis.android.data.local.entity.ArtificialEntity
 import org.openkis.android.data.local.entity.CaveEntity
 import org.openkis.android.data.local.entity.ServerEntity
 import org.openkis.android.data.local.entity.SpringEntity
+import org.openkis.android.data.local.entity.SurveyEntity
 
 @Database(
     entities = [
         CaveEntity::class,
         SpringEntity::class,
         ArtificialEntity::class,
-        ServerEntity::class
+        ServerEntity::class,
+        SurveyEntity::class
     ],
-    version = 3,
+    version = 4,
     exportSchema = false
 )
 abstract class AppDatabase : RoomDatabase() {
@@ -26,4 +29,5 @@ abstract class AppDatabase : RoomDatabase() {
     abstract fun springDao(): SpringDao
     abstract fun artificialDao(): ArtificialDao
     abstract fun serverDao(): ServerDao
+    abstract fun surveyDao(): SurveyDao
 }