+++ app/src/main/java/org/openkis/android/data/local/entity/ArtificialEntity.kt
@@ -24,5 +24,6 @@ data class ArtificialEntity(
     val locality: String = "",
     val photoUrl: String = "",
     val serverUrl: String = "",
+    val dbId: String = "",
     val lastSync: Long = System.currentTimeMillis()
 )